+++ client/Packages/com.beamable/CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Listing content `activeFrom` field no longer resets every time it is viewed in the Inspector.
 - `Json.Serialize` will treat non `IDictionary` values of `IEnumerable` as json arrays.
 - Disabling Content inspectors no longer causes compiler errors.
+- Leaderboard `rankgt` field is not null when specifying outlier.
 
 ### Removed